Add Visual Studio build support + Fixes for 32bit build

+++ src/gfx_macros.h
@@ -78,7 +78,7 @@ typedef struct {
 
 #define gsSPVertex(v, n, v0) \
     GFXCMD_VertexData, \
-    (int64_t)v, n, v0
+    (intptr_t)v, n, v0
 
 #define gsSP2Triangles(v00, v01, v02, flag0, v10, v11, v12, flag1) \
     GFXCMD_Triangle, \
@@ -95,11 +95,11 @@ typedef struct {
 
 #define gsSPDisplayList(dl) \
     GFXCMD_SubDisplayList, \
-    (int64_t)dl
+    (intptr_t)dl
 
 #define gsSPLight(l, n) \
     GFXCMD_Light, \
-    (int64_t)l, n
+    (intptr_t)l, n
 
 #define gsSPTexture(s, t, level, tile, on) \
     GFXCMD_Texture, \
@@ -125,4 +125,4 @@ typedef struct {
 #define gsDPLoadBlock(tile, uls, ult, lrs, dxt) (GFXCMD_None)
 #define gsDPLoadSync() (GFXCMD_None)
 
-typedef int64_t Gfx;
+typedef intptr_t Gfx;
